remove value tampering, dead shaders, stale comments (round 2 consensus)
masked_logit: delete NaN->0 and inf->FLT_MAX sanitization (GPU + CPU).
prio_adv_reduction: delete NaN/inf->0 guard on priority weights.
fp16 scan forward: remove 1e-6 floor on next_state (keep fp16 upper bound).
fp16 scan backward: unify grad_state with fp32 (zero when state <= 0).
delete dead shaders: logcumsumexp fwd/bwd (~70 lines), hgemm_reg (~75 lines),
hgemm_ksplit + reduce_ksplit_fp16 (~93 lines). rename puff_advantage_cuda
to puff_advantage. fix dtype_name "bf16" -> "f16". delete stale comments:
NEON sigmoid, Section 7 TOC, TF32 removal notes, "Float32 only" header.

+++ pufferlib/src/puf_types.h
@@ -98,7 +98,7 @@ struct PufTensor {
     case 1:
       return "i8";
     case 2:
-      return "bf16";
+      return "f16";
     case 4:
       return "f32";
     case 8:
